+++ components/asic/bm1368.c
@@ -119,7 +119,7 @@ void BM1368_set_version_mask(uint32_t version_mask)
     _send_BM1368(TYPE_CMD | GROUP_ALL | CMD_WRITE, version_cmd, 6, BM1368_SERIALTX_DEBUG);
 }
 
-void BM1368_send_hash_frequency(double target_freq) {
+void BM1368_send_hash_frequency(float target_freq) {
     float max_diff = 0.001;
     uint8_t freqbuf[6] = {0x00, 0x08, 0x40, 0xA0, 0x02, 0x41};
     uint8_t postdiv_min = 255;
@@ -169,11 +169,11 @@ void BM1368_send_hash_frequency(double target_freq) {
     current_frequency = target_freq;
 }
 
-bool BM1368_set_frequency(double target_freq) {
+bool BM1368_set_frequency(float target_freq) {
     return do_frequency_transition(target_freq, BM1368_send_hash_frequency, 1368);
 }
 
-static void do_frequency_ramp_up(double target_frequency) {
+static void do_frequency_ramp_up(float target_frequency) {
     ESP_LOGI(TAG, "Ramping up frequency from %.2f MHz to %.2f MHz", current_frequency, target_frequency);
     do_frequency_transition(target_frequency, BM1368_send_hash_frequency, 1368);
 }
