+++ components/asic/bm1370.c
@@ -140,7 +140,7 @@ void BM1370_set_version_mask(uint32_t version_mask)
     _send_BM1370(TYPE_CMD | GROUP_ALL | CMD_WRITE, version_cmd, 6, BM1370_SERIALTX_DEBUG);
 }
 
-void BM1370_send_hash_frequency(double target_freq) {
+void BM1370_send_hash_frequency(float target_freq) {
     // default 200Mhz if it fails
     unsigned char freqbuf[6] = {0x00, 0x08, 0x40, 0xA0, 0x02, 0x41}; // freqbuf - pll0_parameter
     float newf = 200.0;
@@ -209,7 +209,7 @@ static void do_frequency_ramp_up(float target_frequency)
 }
 
 // Add a public function for external use
-bool BM1370_set_frequency(double target_freq)
+bool BM1370_set_frequency(float target_freq)
 {
     return do_frequency_transition(target_freq, BM1370_send_hash_frequency, 1370);
 }
