+++ components/asic/include/bm1397.h
@@ -28,7 +28,7 @@ void BM1397_send_work(bm_job * next_bm_job);
 void BM1397_set_version_mask(uint32_t version_mask);
 int BM1397_set_max_baud(void);
 int BM1397_set_default_baud(void);
-void BM1397_send_hash_frequency(double frequency);
+void BM1397_send_hash_frequency(float frequency);
 task_result * BM1397_process_work();
 
 #endif /* BM1397_H_ */
